feat(package): Rename packages
