+++ src/main/java/com/teamdroid/nikit/MainServerApplication.java
@@ -1,0 +1,13 @@
+package com.teamdroid.nikit;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MainServerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(MainServerApplication.class, args);
+	}
+
+}
