+++ src/main/java/com/teamdroid/nikit/dto/KnowledgeCreateDTO.java
@@ -1,0 +1,9 @@
+package com.teamdroid.nikit.dto;
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class KnowledgeCreateDTO {
+    private String name;
+}
