+++ src/main/java/com/teamdroid/nikit/dto/KnowledgeDTO.java
@@ -1,0 +1,10 @@
+package com.teamdroid.nikit.dto;
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class KnowledgeDTO {
+    private String id;
+    private String name;
+}
