+++ src/main/java/com/teamdroid/nikit/entity/Knowledge.java
@@ -1,0 +1,19 @@
+package com.teamdroid.nikit.entity;
+
+import lombok.*;
+import org.springframework.data.annotation.Id;
+import org.springframework.data.mongodb.core.mapping.Document;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Data
+@Document(collection = "knowledge")
+public class Knowledge {
+
+    @Id
+    private String id;
+    private String name;
+
+}
