+++ src/main/java/com/teamdroid/nikit/entity/User.java
@@ -1,0 +1,14 @@
+package com.teamdroid.nikit.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class User {
+
+    private String username;
+    private String password;
+}
