+++ src/main/java/com/teamdroid/nikit/repository/KnowledgeRepository.java
@@ -1,0 +1,9 @@
+package com.teamdroid.nikit.repository;
+
+import com.teamdroid.nikit.entity.Knowledge;
+import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface KnowledgeRepository extends MongoRepository<Knowledge, String> {
+}
