+++ src/main/java/com/teamdroid/nikit/service/UserService.java
@@ -1,0 +1,20 @@
+package com.teamdroid.nikit.service;
+
+import com.teamdroid.nikit.entity.User;
+import com.teamdroid.nikit.repository.UserRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.Optional;
+
+@Service
+public class UserService {
+
+    @Autowired
+    private UserRepository userRepository;
+
+    public Optional<User> authenticate(String username, String password) {
+        Optional<User> user = userRepository.findByUsername(username);
+        return user.filter(u -> u.getPassword().equals(password));
+    }
+}
