+++ src/main/java/com/teamdroid/nikit/web/WebController.java
@@ -1,0 +1,13 @@
+package com.teamdroid.nikit.web;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class WebController {
+
+    @GetMapping("/")
+    public String index() {
+        return "index";
+    }
+}
