--- src/main/java/com/teamdroid/wise/MainServerApplication.java
@@ -1,13 +1,0 @@
-package com.teamdroid.wise;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MainServerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(MainServerApplication.class, args);
-	}
-
-}
