--- src/main/java/com/teamdroid/wise/dto/KnowledgeCreateDTO.java
@@ -1,9 +1,0 @@
-package com.teamdroid.wise.dto;
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class KnowledgeCreateDTO {
-    private String name;
-}
