--- src/main/java/com/teamdroid/wise/dto/KnowledgeDTO.java
@@ -1,10 +1,0 @@
-package com.teamdroid.wise.dto;
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class KnowledgeDTO {
-    private String id;
-    private String name;
-}
