--- src/main/java/com/teamdroid/wise/entity/SessionToken.java
@@ -1,16 +1,0 @@
-package com.teamdroid.wise.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-import java.util.Date;
-
-/**
- * Representa un token de sesión, incluyendo el token JWT y su fecha de expiración.
- */
-@Data
-@AllArgsConstructor
-public class SessionToken {
-    private String token;
-    private Date expiresAt;
-}
