--- src/main/java/com/teamdroid/wise/entity/User.java
@@ -1,14 +1,0 @@
-package com.teamdroid.wise.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class User {
-
-    private String username;
-    private String password;
-}
