--- src/main/java/com/teamdroid/wise/service/UserService.java
@@ -1,20 +1,0 @@
-package com.teamdroid.wise.service;
-
-import com.teamdroid.wise.entity.User;
-import com.teamdroid.wise.repository.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import java.util.Optional;
-
-@Service
-public class UserService {
-
-    @Autowired
-    private UserRepository userRepository;
-
-    public Optional<User> authenticate(String username, String password) {
-        Optional<User> user = userRepository.findByUsername(username);
-        return user.filter(u -> u.getPassword().equals(password));
-    }
-}
