--- src/main/java/com/teamdroid/wise/web/WebController.java
@@ -1,13 +1,0 @@
-package com.teamdroid.wise.web;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class WebController {
-
-    @GetMapping("/")
-    public String index() {
-        return "index";
-    }
-}
