+++ src/test/java/com/teamdroid/nikit/MainServerApplicationTests.java
@@ -1,0 +1,13 @@
+package com.teamdroid.nikit;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class MainServerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
