--- src/test/java/com/teamdroid/wise/MainServerApplicationTests.java
@@ -1,13 +1,0 @@
-package com.teamdroid.wise;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MainServerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
